# Revive JMH benchmarks and improve circular buffer next index calculation