--- a/build.gradle
+++ b/build.gradle
@@ -3,6 +3,7 @@ plugins {
     id 'org.asciidoctor.convert' version '1.6.0'
     id "org.gradle.test-retry" version "1.2.1"
     id "io.github.gradle-nexus.publish-plugin"  version "1.1.0"
+    id "me.champeau.jmh" version "0.6.5"
     id 'idea'
 }
 
@@ -41,6 +42,7 @@ configure(project.coreProjects) {
     apply plugin: 'signing'
     apply from: "${rootDir}/publishing.gradle"
     apply plugin: "org.gradle.test-retry"
+    apply plugin: 'me.champeau.jmh'
 
     tasks.withType(Test).all {
         if (JavaVersion.current().isCompatibleWith(JavaVersion.VERSION_13)) {