--- a/resilience4j-circularbuffer/src/jmh/resources/benchmark_results.txt
+++ b/resilience4j-circularbuffer/src/jmh/resources/benchmark_results.txt
@@ -1,13 +1,28 @@
-Benchmark                                                       Mode  Cnt    Score    Error  Units
-CircularBufferBenchmark.circularBuffer                          avgt   20  135.347 ± 20.444  ns/op
-CircularBufferBenchmark.circularBuffer:circularBufferAddEvent   avgt   20  267.520 ± 55.244  ns/op
-CircularBufferBenchmark.circularBuffer:circularBufferSize       avgt   20    9.401 ±  4.046  ns/op
-CircularBufferBenchmark.circularBuffer:circularBufferTakeEvent  avgt   20   75.721 ± 29.743  ns/op
-CircularBufferBenchmark.circularBuffer:circularBufferToList     avgt   20  188.747 ± 72.451  ns/op
+Before
 
-Benchmark                                                                             Mode  Cnt    Score    Error  Units
-ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue                              avgt   20  116.028 ± 14.322  ns/op
-ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueAdd   avgt   20  179.901 ± 35.468  ns/op
-ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePeek  avgt   20   90.970 ± 30.847  ns/op
-ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePoll  avgt   20  123.493 ± 20.239  ns/op
-ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueSize  avgt   20    5.875 ±  1.569  ns/op
+Benchmark                                                                             Mode  Cnt     Score     Error  Units
+CircularBufferBenchmark.circularBuffer                                                avgt   20   938.412 ± 178.388  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferAddEvent                         avgt   20   253.779 ±  32.676  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferSize                             avgt   20     2.554 ±   0.168  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferTakeEvent                        avgt   20   111.113 ±  31.443  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferToList                           avgt   20  3386.199 ± 777.348  ns/op
+
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue                              avgt   20   251.910 ±   7.112  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueAdd   avgt   20   413.260 ±   9.749  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePeek  avgt   20    29.005 ±   0.723  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePoll  avgt   20   400.014 ±  15.411  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueSize  avgt   20     4.011 ±   0.052  ns/op
+
+After
+Benchmark                                                                             Mode  Cnt     Score      Error  Units
+CircularBufferBenchmark.circularBuffer                                                avgt   20  1323.005 ±  268.516  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferAddEvent                         avgt   20   185.087 ±    8.428  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferSize                             avgt   20     2.285 ±    0.017  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferTakeEvent                        avgt   20    82.474 ±    2.903  ns/op
+CircularBufferBenchmark.circularBuffer:circularBufferToList                           avgt   20  5022.176 ± 1068.645  ns/op
+
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue                              avgt   20   129.095 ±   10.612  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueAdd   avgt   20   253.528 ±   24.446  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePeek  avgt   20    29.658 ±    0.128  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueuePoll  avgt   20   106.226 ±    4.159  ns/op
+ConcurrentEvictingQueueBenchmark.concurrentEvictingQueue:concurrentEvictingQueueSize  avgt   20     2.534 ±    0.210  ns/op